+++ lib/tasks/index.js
@@ -17,7 +17,7 @@ module.exports = {
   coveralls : require('./coveralls' )($, gulp, paths),
   lint : require('./lint' )($, gulp, paths),
   karma : require('./karma' )($, gulp, paths),
-  // mocha : require('./mocha' )($, gulp, paths),
+  mocha : require('./mocha' )($, gulp, paths),
 }
