+++ lib/tasks/karma.js
@@ -12,7 +12,7 @@ module.exports = function($, gulp, paths){
     return karma.start({
       configFile: paths.spec.root + '/karma.conf.js',
       singleRun: true
-    });
+    }, done);
   };
 
 };
