+++ src/spec/server/test.js
@@ -1,3 +1,7 @@
+/*jshint -W079 */
+
+var expect = require('chai').expect;
+
 describe('it should test something', function() {
   it('should work', function() {
     expect(true).to.equal(true);
